refactor(view): view 패키지 생성 및 구조 변경
- InputView, OutputView 클래스를 view 패키지로 이동
